Add CSP_chanSetAsync to create an async channel.
For an async channel, writing does not block the thread. This is
especially useful when you have a very fast writer, or interrupt handler
that may not block.

If a chanOut happens for the second time, while the value was not read
yet, the buffer is overwritten. If you want all values, you have to make
sure the reader is fast enough, or values will be dropped.